Change all the things

+++ src/index.ts
@@ -1,12 +1,12 @@
-#!/usr/bin/env node
-let fs = require("fs-extra");
-let path = require("path");
-let chalk = require("chalk");
-let semver = require("semver");
-let meow = require("meow");
-let getWorkspaces = require("get-workspaces").default;
+import fs from "fs-extra";
+import path from "path";
+import chalk from "chalk";
+import semver from "semver";
+import meow from "meow";
+import getWorkspaces from "get-workspaces";
+import { DEPENDENCY_TYPES, Workspace } from "./constants";
+import upgrade from "./upgrade";
 
-let cmd = process.argv[2];
 const { input } = meow(
   `
     Usage
@@ -14,13 +14,14 @@ const { input } = meow(
     Commands
       link         link a foreign monorepo to the local monorepo
       unlink       unlink foreign packages
+      upgrade      upgrade versions of packages from a foreign monorepo in the local monorepo
     `,
   {}
 );
 
-if (input[0] !== "link" && input[0] !== "unlink") {
+if (input[0] !== "link" && input[0] !== "unlink" && input[0] !== "upgrade") {
   console.error(
-    `link-monorepos-badly only supports the \`link\` and \`unlink\` commands, not ${
+    `link-monorepos-badly only supports the \`link\`, \`unlink\` and \`upgrade\` commands, not ${
       input[0]
     } `
   );
@@ -32,10 +33,8 @@ let mode = input[0];
 let foreignMonorepo = path.join(process.cwd(), input[1]);
 let localMonorepo = process.cwd();
 
-let depTypes = ["dependencies", "peerDependencies"];
-
 (async () => {
-  let [foreignWorkspaces, localWorkspaces] = await Promise.all([
+  let [_foreignWorkspaces, _localWorkspaces] = await Promise.all([
     getWorkspaces({
       tools: ["bolt", "yarn", "root"],
       cwd: foreignMonorepo
@@ -46,7 +45,30 @@ let depTypes = ["dependencies", "peerDependencies"];
     })
   ]);
 
-  let foreignWorkspacesMap = new Map();
+  if (_foreignWorkspaces === null || _localWorkspaces === null) {
+    throw new Error("Workspaces could not be found");
+  }
+
+  let foreignWorkspaces = _foreignWorkspaces as Workspace[];
+  let localWorkspaces = _localWorkspaces as Workspace[];
+
+  let localRootPkgJson = require(path.join(localMonorepo, "package.json"));
+
+  let localRootWorkspace: Workspace = {
+    name: localRootPkgJson.name,
+    dir: localMonorepo,
+    config: localRootPkgJson
+  };
+
+  if (mode === "upgrade") {
+    await upgrade(
+      localWorkspaces.concat(localRootWorkspace),
+      foreignWorkspaces
+    );
+    return;
+  }
+
+  let foreignWorkspacesMap = new Map<string, Workspace>();
 
   for (let foreignWorkspace of foreignWorkspaces) {
     foreignWorkspacesMap.set(foreignWorkspace.name, foreignWorkspace);
@@ -55,11 +77,15 @@ let depTypes = ["dependencies", "peerDependencies"];
   await Promise.all(
     localWorkspaces.map(workspace => {
       let depsToLink = new Map();
-      for (let depType of depTypes) {
+      for (let depType of DEPENDENCY_TYPES) {
         if (workspace.config[depType]) {
           for (let depName in workspace.config[depType]) {
             if (foreignWorkspacesMap.has(depName)) {
-              depsToLink.set(depName, workspace.config[depType][depName]);
+              depsToLink.set(
+                depName,
+                // @ts-ignore
+                workspace.config[depType][depName]
+              );
             }
           }
         }
@@ -72,23 +98,23 @@ let depTypes = ["dependencies", "peerDependencies"];
             "node_modules",
             depName
           );
-          await fs.remove(nodeModulesDepDir);
+          if (mode === "link" || mode === "unlink") {
+            await fs.remove(nodeModulesDepDir);
+          }
           if (mode === "link") {
-            await fs.ensureSymlink(
-              foreignWorkspacesMap.get(depName).dir,
-              nodeModulesDepDir
-            );
+            let foreignWorkspace = foreignWorkspacesMap.get(
+              depName
+            ) as Workspace;
+
+            await fs.ensureSymlink(foreignWorkspace.dir, nodeModulesDepDir);
             if (
-              !semver.satisfies(
-                foreignWorkspacesMap.get(depName).config.version,
-                depVersion
-              )
+              !semver.satisfies(foreignWorkspace.config.version, depVersion)
             ) {
               console.error(
                 `⚠️ ${chalk.green(workspace.name)} depends on ${chalk.red(
                   depName + "@" + depVersion
                 )} but ${chalk.green(depName)} is at ${chalk.green(
-                  foreignWorkspacesMap.get(depName).config.version
+                  foreignWorkspace.config.version
                 )} in the foreign monorepo`
               );
             }
@@ -97,7 +123,7 @@ let depTypes = ["dependencies", "peerDependencies"];
                 workspace.name
               )}`
             );
-          } else {
+          } else if (mode === "unlink") {
             console.log(
               `🎉 unlinked ${chalk.green(depName)} from ${chalk.green(
                 workspace.name
